+++ packages/featureserver/CHANGELOG.md
@@ -2,11 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [4.0.0] - 06-14-2022
 ### Added
 - override default currentVersion and fullVersion via app.locals.config.featureServer
 - override default service-description and layer-description via app.locals.config.featureServer
 
+### Changed
+- function signature on the restInfo method
 ## [3.5.1] - 06-01-2022
 ### Changed
 - bump winnow version
@@ -485,6 +487,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.0.0] - 2017-01-25
 * Initial Release
 
+[4.0.0]: https://github.com/koopjs/featureserver/compare/v3.5.0...v4.0.0
 [3.5.0]: https://github.com/koopjs/featureserver/compare/v3.4.0...v3.5.0
 [3.4.0]: https://github.com/koopjs/featureserver/compare/v3.3.4...v3.4.0
 [3.3.4]: https://github.com/koopjs/featureserver/compare/v3.3.3...v3.3.4